mock-backend/requirements.txt
fastapi>=0.110,<1.0
pydantic>=2.4,<3.0
uvicorn[standard]>=0.27,<0.30
python-dotenv>=1.0,<2.0
